requirements.txt
minio
python-dotenv
tqdm